create new module/file for threads
