+++ backend/app/app/api/api_v1/endpoints/__init__.py
@@ -1,0 +1,11 @@
+from .login import router
+from .users import router
+from .utils import router
+from .ue_movement import router
+from .paths import router
+from .gNB import router
+from .Cell import router
+from .UE import router
+from .qosInformation import router
+from .qosMonitoring import router
+from .monitoringevent import router
